+++ src/main/java/minecraftarmorweapon/procedures/GomanorikenBlockburotukugaYoukuritukusaretatokiProcedure.java
@@ -15,6 +15,6 @@ public class GomanorikenBlockburotukugaYoukuritukusaretatokiProcedure {
 		if (world instanceof ServerLevel _level)
 			_level.getServer().getCommands().performPrefixedCommand(new CommandSourceStack(CommandSource.NULL, new Vec3(x, y, z), Vec2.ZERO, _level, 4, "", Component.literal(""), _level.getServer(), null).withSuppressedOutput(),
 					"give @p minecraft_armor_weapon:gomanoriken{Enchantments:[{id:smite,lvl:5}]}");
-		world.setBlock(new BlockPos(x, y, z), Blocks.AIR.defaultBlockState(), 3);
+		world.setBlock(new BlockPos((int) (x), (int) (y), (int) (z)), Blocks.AIR.defaultBlockState(), 3);
 	}
 }
