+++ src/main/java/minecraftarmorweapon/procedures/ResetMaxYoukuritukusitatokiProcedure.java
@@ -34,7 +34,7 @@ public class ResetMaxYoukuritukusitatokiProcedure {
 		entity.setSecondsOnFire(0);
 		if (world instanceof Level _level) {
 			if (!_level.isClientSide()) {
-				_level.playSound(null, new BlockPos(x, y, z), ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("block.anvil.place")), SoundSource.PLAYERS, 1, 1);
+				_level.playSound(null, new BlockPos((int) (x), (int) (y), (int) (z)), ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("block.anvil.place")), SoundSource.PLAYERS, 1, 1);
 			} else {
 				_level.playLocalSound(x, y, z, ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("block.anvil.place")), SoundSource.PLAYERS, 1, 1, false);
 			}
