+++ src/main/java/minecraftarmorweapon/procedures/TestBlockburotukugaYoukuritukusaretatokiProcedure.java
@@ -14,7 +14,7 @@ public class TestBlockburotukugaYoukuritukusaretatokiProcedure {
 		if (world instanceof ServerLevel _serverworld) {
 			StructureTemplate template = _serverworld.getStructureManager().getOrCreate(new ResourceLocation("minecraft_armor_weapon", "kenisekiskelton"));
 			if (template != null) {
-				template.placeInWorld(_serverworld, new BlockPos(x, y, z), new BlockPos(x, y, z), new StructurePlaceSettings().setRotation(Rotation.NONE).setMirror(Mirror.NONE).setIgnoreEntities(false), _serverworld.random, 3);
+				template.placeInWorld(_serverworld, new BlockPos((int) (x), (int) (y), (int) (z)), new BlockPos((int) (x), (int) (y), (int) (z)), new StructurePlaceSettings().setRotation(Rotation.NONE).setMirror(Mirror.NONE).setIgnoreEntities(false), _serverworld.random, 3);
 			}
 		}
 	}
